Take the Internet out of the loop
We had been pinging a variety of sites across the Internet. This meant
that any errors we get could be local to our gear or they could be
remote. We had no way of knowing.

Take the Internet out of the loop. Ping the first named router our
traffic goes through after leaving our gear.